requirements.txt
ctypes
time
sys
random
keyboard
os
json
heapq
math
win32gui
pillow